% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyprojectAZ.R
\name{tidyprojectAZ_tests}
\alias{tidyprojectAZ_tests}
\title{az specific tests}
\usage{
tidyprojectAZ_tests(proj_name = getwd(), silent = FALSE)
}
\arguments{
\item{proj_name}{character. default = current working directory. path to directory.}

\item{silent}{logical. default = FALSE. suppress messages or not}
}
\description{
Not intended to be run directly by user
}
